Fix missing Dminus4 terms

* moved comments

* soft script works

* virt new line EOF; changed dd soft-fractioned

* put negative |ME|^2 info into debug

* more explicit generation of soft; to be sure

* reduced test precision

* we want to check for inf

* separate Helicity and Pol calculations

* tweaked readme

* Merge remote-tracking branch 'origin/restructure-code' into fix-missing-Dminus4-terms

+++ include/XSections/XSection_Virt.hpp
@@ -8,7 +8,7 @@ public:
    XSection_Virt(
       XSectionParameters const& parameters,
       double m1, double m2,
-      std::function<double(double, double, double, double, double, int, double)> f_,
+      std::function<double(double, double, double, double, int, int, double)> f_,
       std::vector<std::array<int, 3>> flav,
       int integration_precision, int integration_verbosity
    ) : XSection{parameters, m1, m2, flav, integration_precision, integration_verbosity}, f{f_}
@@ -18,7 +18,7 @@ public:
    double integrand(const double xx[]);
 
 private:
-   std::function<double(double, double, double, double, double, int, double)> f;
+   std::function<double(double, double, double, double, int, int, double)> f;
 };
 
 #endif // XSECTION_VIRT_H_
